requirements.txt
web3==6.15.1
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.4
